+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.8
 
+- Syntax `A[i...]`, with `A` a uniform array and `i...` sub-array indices,
+  yields a uniform array (or a scalar).
+
 - New non-exported method `StructuredArrays.parameterless(T)` which yields type
   `T` without parameter specifications. For example:
   `StructuredArrays.parameterless(Vector{Float32})` yields `Array`.